refactor(persistence): ADR-0035 Stage 2.5 C2 slice 1 — extract kirra-ota-campaign
Relocate the pure OTA governor-artifact campaign engine (src/ota_campaign.rs)
into a new lean crate, kirra-ota-campaign, leaving src/ota_campaign.rs as a
`pub use kirra_ota_campaign::*;` re-export shim. This clears the last C2
domain-type coupling in verifier_store::ota_campaigns (it named
crate::ota_campaign::{Campaign, CampaignState, HaltReason, NodeArtifactStatus,
AdvanceOutcome}) so a later kirra-persistence extraction can name the campaign
contract without the state machine living inside the verifier service tree.

The engine is genuinely pure: its only non-serde code deps are
kirra_core::FleetPosture (the fail-closed advance gate) and
kirra_release_token::uptane::UptaneMetadataSet (authoring-time digest
authorization) — both already lean, verifier-independent crates, so no cycle.
The one internal import (crate::verifier::FleetPosture, itself a re-export of
kirra_core::FleetPosture) becomes a direct kirra_core::FleetPosture; a broken
intra-doc link to crate::attestation is demoted to a plain code span.

Behaviour-preserving: the module content moves verbatim, every existing
crate::ota_campaign::* / kirra_verifier::ota_campaign::* path (store, service
binary, campaign_monitor, metrics, tests) resolves through the shim unchanged,
and the 27 engine unit tests move with it (now run under
`cargo test -p kirra-ota-campaign` / workspace CI).

Mirrors the kirra-fleet-types idiom used for the federation types.

Verified: workspace build + `cargo test --lib` (758, = 785 − the 27 moved) +
power-loss drill (audit_chain_prefix_on_kill, 3) + kirra-ota-campaign (27) +
fmt + quality guardrails + orphan-cores all green.
